tests/requirements.txt
pytest
requests
urllib3>=1.26,<2
Pillow
numpy
